pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vibevoice"
version = "0.0.1"
authors = [
  { name="vibevoice team", email="vibepod@microsoft.com" },
]
description = "A model for speech generation with an AR + diffusion architecture."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    # "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "torch",
    "accelerate==1.6.0",
    "transformers==4.51.3", # we develop this project on transformers==4.51.3, later version may not be compatible
    "datasets==3.5.0", # avoid issues with torchcodec
    "peft", # for finetuning
    "llvmlite>=0.40.0",
    "numba>=0.57.0",
    "diffusers",
    "tqdm",
    "numpy",
    "scipy",
    "librosa",
    "soundfile",
    "ml-collections",
    "absl-py",
    "gradio",
    "av",
    "aiortc",
    # API server deps
    "fastapi",
    "uvicorn[standard]"
    ,
    "prometheus-client"
    ,
    "pyyaml"
]


[project.urls]
"Homepage" = "https://github.com/microsoft/VibeVoice"
"Bug Tracker" = "https://github.com/microsoft/VibeVoice/issues"

[tool.setuptools.packages.find]
where = ["."]
